把消息处理内变为内部的。
