+++ src/main/java/ice/games/snake/SnakeGame.java
@@ -32,4 +32,19 @@ public class SnakeGame {
 				.getReader().readLine());
 	}
 
+	private static class MessageListener {
+
+		static void onMessage(AtmosphereResource resource, String message) {
+			Snake snake = (Snake) resource.session().getAttribute("snake");
+			if ("left".equals(message)) {
+				snake.setDirection(Direction.LEFT);
+			} else if ("up".equals(message)) {
+				snake.setDirection(Direction.UP);
+			} else if ("right".equals(message)) {
+				snake.setDirection(Direction.RIGHT);
+			} else if ("down".equals(message)) {
+				snake.setDirection(Direction.DOWN);
+			}
+		}
+	}
 }
